+++ src/Symfony/Component/Form/Util/InheritDataAwareIterator.php
@@ -19,11 +19,8 @@ namespace Symfony\Component\Form\Util;
  * the form and traverses its children as well.
  *
  * @author Bernhard Schussek <bschussek@gmail.com>
- *
- * @deprecated Deprecated since version 2.2, to be removed in 2.3. Use
- *             {@link InheritDataAwareIterator} instead.
  */
-class VirtualFormAwareIterator extends \ArrayIterator implements \RecursiveIterator
+class InheritDataAwareIterator extends VirtualFormAwareIterator
 {
     /**
      * Creates a new iterator.
@@ -32,18 +29,7 @@ class VirtualFormAwareIterator extends \ArrayIterator implements \RecursiveItera
      */
     public function __construct(array $forms)
     {
-        trigger_error('VirtualFormAwareIterator is deprecated since version 2.2 and will be removed in 2.3. Use InheritDataAwareIterator instead.', E_USER_DEPRECATED);
-
-        parent::__construct($forms);
-    }
-
-    public function getChildren()
-    {
-        return new static($this->current()->all());
-    }
-
-    public function hasChildren()
-    {
-        return $this->current()->getConfig()->getInheritData();
+        // Skip the deprecation error
+        \ArrayIterator::__construct($forms);
     }
 }