+++ src/Symfony/Component/Form/CHANGELOG.md
@@ -10,6 +10,9 @@ CHANGELOG
  * deprecated passing a Request instance to FormInterface::bind()
  * added options "method" and "action" to FormType
  * deprecated option "virtual", renamed it to "inherit_data"
+ * deprecated VirtualFormAwareIterator in favor of InheritDataAwareIterator
+ * [BC BREAK] removed the "array" type hint from DataMapperInterface
+ * improved forms inheriting their parent data to actually return that data from getData(), getNormData() and getViewData()
 
 2.2.0
 -----